add scan operator
