+++ pixels-turbo/pixels-worker-common/src/main/java/io/pixelsdb/pixels/worker/common/BasePartitionWorker.java
@@ -165,7 +165,7 @@ public class BasePartitionWorker extends Worker<PartitionInput, PartitionOutput>
                     hashValues.add(hash);
                 }
             }
-            partitionOutput.setPath(outputPath);
+            partitionOutput.addOutput(outputPath);
             partitionOutput.setHashValues(hashValues);
 
             pixelsWriter.close();
